// generated/BathToken/BathToken.ts
export class Approval extends ethereum.Event {
  get params(): Approval__Params {
    return new Approval__Params(this);
  }
}

export class Approval__Params {
  _event: Approval;

  constructor(event: Approval) {
    this._event = event;
  }

  get owner(): Address {
    return this._event.parameters[0].value.toAddress();
  }

  get spender(): Address {
    return this._event.parameters[1].value.toAddress();
  }

  get value(): BigInt {
    return this._event.parameters[2].value.toBigInt();
  }
}

export class Deposit extends ethereum.Event {
  get params(): Deposit__Params {
    return new Deposit__Params(this);
  }
}

export class Deposit__Params {
  _event: Deposit;

  constructor(event: Deposit) {
    this._event = event;
  }

  get depositedAmt(): BigInt {
    return this._event.parameters[0].value.toBigInt();
  }

  get asset(): Address {
    return this._event.parameters[1].value.toAddress();
  }

  get sharesReceived(): BigInt {
    return this._event.parameters[2].value.toBigInt();
  }

  get depositor(): Address {
    return this._event.parameters[3].value.toAddress();
  }
}

export class LogInit extends ethereum.Event {
  get params(): LogInit__Params {
    return new LogInit__Params(this);
  }
}

export class LogInit__Params {
  _event: LogInit;

  constructor(event: LogInit) {
    this._event = event;
  }

  get timeOfInit(): BigInt {
    return this._event.parameters[0].value.toBigInt();
  }
}

export class Transfer extends ethereum.Event {
  get params(): Transfer__Params {
    return new Transfer__Params(this);
  }
}

export class Transfer__Params {
  _event: Transfer;

  constructor(event: Transfer) {
    this._event = event;
  }

  get from(): Address {
    return this._event.parameters[0].value.toAddress();
  }

  get to(): Address {
    return this._event.parameters[1].value.toAddress();
  }

  get value(): BigInt {
    return this._event.parameters[2].value.toBigInt();
  }
}

export class Withdraw extends ethereum.Event {
  get params(): Withdraw__Params {
    return new Withdraw__Params(this);
  }
}

export class Withdraw__Params {
  _event: Withdraw;

  constructor(event: Withdraw) {
    this._event = event;
  }

  get amountWithdrawn(): BigInt {
    return this._event.parameters[0].value.toBigInt();
  }

  get asset(): Address {
    return this._event.parameters[1].value.toAddress();
  }

  get sharesWithdrawn(): BigInt {
    return this._event.parameters[2].value.toBigInt();
  }

  get withdrawer(): Address {
    return this._event.parameters[3].value.toAddress();
  }

  get fee(): BigInt {
    return this._event.parameters[4].value.toBigInt();
  }

  get feeTo(): Address {
    return this._event.parameters[5].value.toAddress();
  }
}